+++ build.gradle
@@ -39,5 +39,5 @@ run {
 
 application {
     // Change this to your main class.
-    mainClassName = "Duke"
+    mainClassName = "duke.Duke"
 }
