+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,7 +1,16 @@
-public class DoneCommand extends Command {
+package duke.command;
+
+import duke.DukeException;
+import duke.common.Message;
+import duke.storage.Storage;
+import duke.task.Task;
+import duke.task.TaskList;
+import duke.ui.UserInterface;
+
+public class DeleteCommand extends Command {
     private int index;
 
-    public DoneCommand(int index) {
+    public DeleteCommand(int index) {
         this.index = index;
     }
 
@@ -10,8 +19,9 @@ public class DoneCommand extends Command {
         if (index <= 0 || index > taskList.size()) {
             throw new DukeException(Message.MESSAGE_INVALID_TASK_INDEX);
         }
-        Task task = taskList.markDone(index);
+        Task task = taskList.delete(index);
         storage.save(taskList.getSimplifiedTaskRepresentations());
-        ui.showMarkDone(task);
+        ui.showDeletion(task);
+        ui.showTaskSize(taskList);
     }
 }
