+++ src/main/java/duke/common/Message.java
@@ -1,3 +1,5 @@
+package duke.common;
+
 public class Message {
 
     public static final String MESSAGE_WELCOME = "Hello! I'm Duke\n%1$sWhat can I do for you?\n%2$s";
@@ -14,6 +16,7 @@ public class Message {
             "Events must have /at and cannot be empty. E.g. event project meeting /at 2/12/2019 1800";
     public static final String MESSAGE_INVALID_TASK_INDEX = "Invalid task number. Task numbers follow 1-based indexing.";
     public static final String MESSAGE_ERROR_MISSING_STORAGE_FILE = "%1$sMissing storage file: %2$s";
+    public static final String MESSAGE_ERROR_CREATING_STORAGE_FILE = "Unable to create new file %1$s. %2$s";
     public static final String MESSAGE_STORAGE_FILE_CREATED = "%1$sCreated new empty storage file: %2$s";
     public static final String MESSAGE_ERROR_READING_FROM_FILE = "Error when reading file";
     public static final String MESSAGE_INVALID_DATE_FORMAT = "Invalid date. Follow d/M/yyyy HHmm format. E.g. 2/12/2019 1800";
