+++ src/main/java/duke/parser/Parser.java
@@ -1,3 +1,12 @@
+package duke.parser;
+
+import duke.DukeException;
+import duke.command.*;
+import duke.common.Message;
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Todo;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
