+++ src/main/java/duke/storage/Storage.java
@@ -1,3 +1,14 @@
+package duke.storage;
+
+import static duke.common.Message.*;
+
+import duke.DukeException;
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.Todo;
+import duke.ui.UserInterface;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
@@ -7,6 +18,7 @@ import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
+
 
 public class Storage {
     private final String filePath;
@@ -63,11 +75,11 @@ public class Storage {
                     task = new Event(description, at);
                     break;
                 default:
-                    throw new DukeException(Message.MESSAGE_ERROR_READING_FROM_FILE);
+                    throw new DukeException(MESSAGE_ERROR_READING_FROM_FILE);
             }
             task.setDone(doneStatus.equals("1"));
         } catch (ArrayIndexOutOfBoundsException e) {
-            throw new DukeException(Message.MESSAGE_ERROR_READING_FROM_FILE);
+            throw new DukeException(MESSAGE_ERROR_READING_FROM_FILE);
         }
         return task;
     }
@@ -77,12 +89,12 @@ public class Storage {
         if (storageFile.exists()) {
             return;
         }
-        System.out.println(String.format(Message.MESSAGE_ERROR_MISSING_STORAGE_FILE, UserInterface.LINE_PREFIX, filePath));
+        System.out.println(String.format(MESSAGE_ERROR_MISSING_STORAGE_FILE, UserInterface.LINE_PREFIX, filePath));
         try {
             storageFile.createNewFile();
-            System.out.println(String.format(Message.MESSAGE_STORAGE_FILE_CREATED, UserInterface.LINE_PREFIX,filePath));
+            System.out.println(String.format(MESSAGE_STORAGE_FILE_CREATED, UserInterface.LINE_PREFIX,filePath));
         } catch (IOException ioe) {
-            throw new DukeException(String.format("Unable to create new file %1$s. %2$s", filePath, ioe.getMessage()));
+            throw new DukeException(String.format(MESSAGE_ERROR_CREATING_STORAGE_FILE, filePath, ioe.getMessage()));
         }
     }
 
@@ -94,7 +106,7 @@ public class Storage {
             }
             fileWriter.close();
         } catch (IOException ioe) {
-            throw new DukeException(Message.MESSAGE_ERROR_MISSING_STORAGE_FILE);
+            throw new DukeException(MESSAGE_ERROR_MISSING_STORAGE_FILE);
         }
     }
 
@@ -104,7 +116,7 @@ public class Storage {
             fileWriter.write(simplifiedTaskRepresentation + "\n");
             fileWriter.close();
         } catch (IOException ioe) {
-            throw new DukeException(Message.MESSAGE_ERROR_MISSING_STORAGE_FILE);
+            throw new DukeException(MESSAGE_ERROR_MISSING_STORAGE_FILE);
         }
     }
 }
